Split alloc tests into their own binary
