+++ firmware/qemu/src/bin/alloc.release.out
@@ -1,0 +1,10 @@
+0.000000 INFO Box<u32>: 42
+0.000001 INFO Box<Box<u32>>: 1337
+0.000002 INFO in nested 123
+0.000003 INFO Box<NestedStruct>: NestedStruct { a: 170, b: 305419896 }
+0.000004 INFO Rc<u32>: 42
+0.000005 INFO Arc<u32>: 42
+0.000006 INFO Vec<u32>: [1, 2, 3, 4]
+0.000007 INFO Vec<i32>: [-1, 2, 3, 4]
+0.000008 INFO Vec<Box<i32>>: [-1, 2, 3, 4]
+0.000009 INFO Box<Vec<i32>>: [-1, 2, 3, 4]
