+++ firmware/qemu/src/bin/log.rs
@@ -1,6 +1,5 @@
 #![no_std]
 #![no_main]
-#![cfg_attr(feature = "alloc", feature(alloc_error_handler))]
 
 use core::sync::atomic::{AtomicU32, Ordering};
 use cortex_m_rt::entry;
@@ -10,14 +9,8 @@ use defmt::Format;
 use defmt_semihosting as _; // global logger
 use panic_probe as _; // panicking behavior
 
-#[cfg(feature = "alloc")]
-extern crate alloc;
-
 #[entry]
 fn main() -> ! {
-    #[cfg(feature = "alloc")]
-    if_alloc::init();
-
     defmt::info!("Hello!");
     defmt::info!("World!");
     defmt::info!("The answer is {:u8}", 42);
@@ -373,32 +366,6 @@ fn main() -> ! {
         [(1u32, 2u32), (3, 4), (5, 6), (7, 8)]
     );
 
-    #[cfg(feature = "alloc")]
-    {
-        use alloc::boxed::Box;
-        use alloc::rc::Rc;
-        use alloc::string::String;
-        use alloc::sync::Arc;
-        use alloc::vec;
-
-        defmt::info!("Box<u32>: {:?}", Box::new(42u32));
-        defmt::info!("Box<Box<u32>>: {:?}", Box::new(Box::new(1337u32)));
-        defmt::info!("Box<NestedStruct>: {:?}", Box::new(nested()));
-        defmt::info!("Rc<u32>: {:?}", Rc::new(42u32));
-        defmt::info!("Arc<u32>: {:?}", Arc::new(42u32));
-        defmt::info!("Vec<u32>: {:?}", vec![1u32, 2, 3, 4]);
-        defmt::info!("Vec<i32>: {:?}", vec![-1i32, 2, 3, 4]);
-        defmt::info!(
-            "Vec<Box<i32>>: {:?}",
-            vec![Box::new(-1i32), Box::new(2), Box::new(3), Box::new(4)]
-        );
-        defmt::info!("Box<Vec<i32>>: {:?}", Box::new(vec![-1i32, 2, 3, 4]));
-        defmt::info!(
-            "String: {:?}",
-            String::from("Hello! I'm a heap-allocated String")
-        );
-    }
-
     loop {
         debug::exit(debug::EXIT_SUCCESS)
     }
@@ -423,25 +390,4 @@ fn timestamp() -> u64 {
     // monotonic counter
     static COUNT: AtomicU32 = AtomicU32::new(0);
     COUNT.fetch_add(1, Ordering::Relaxed) as u64
-}
-
-#[cfg(feature = "alloc")]
-mod if_alloc {
-    use alloc_cortex_m::CortexMHeap;
-    use core::alloc::Layout;
-
-    #[global_allocator]
-    static ALLOCATOR: CortexMHeap = CortexMHeap::empty();
-
-    #[alloc_error_handler]
-    fn oom(_: Layout) -> ! {
-        loop {}
-    }
-
-    pub fn init() {
-        // Initialize the allocator BEFORE you use it
-        let start = cortex_m_rt::heap_start() as usize;
-        let size = 1024; // in bytes
-        unsafe { ALLOCATOR.init(start, size) }
-    }
 }
