+++ firmware/qemu/test.sh
@@ -1,12 +1,16 @@
 #!/usr/bin/env bash
 
-set -e
+set -o errexit
 
-FEATURES=${FEATURES-}
+function test() {
+    local bin=$1
+    local features=${2-,}
+    
+    cargo rb "$bin" --features "$features" | tee "src/bin/$bin.out.new" | diff "src/bin/$bin.out" -
+    cargo rrb "$bin" --features "$features" | tee "src/bin/$bin.release.out.new" | diff "src/bin/$bin.release.out" -
+}
 
-for testfile in src/bin/*.rs; do
-    testname=$(basename "$testfile" .rs)
-
-    cargo rb "$testname" --features "$FEATURES" | tee "src/bin/$testname.out.new" | diff "src/bin/$testname.out" -
-    cargo rrb "$testname" --features "$FEATURES" | tee "src/bin/$testname.release.out.new" | diff "src/bin/$testname.release.out" -
-done
+test "log"
+if rustc -V | grep nightly; then
+    test "alloc" "alloc"
+fi
